// apps/frontend/src/api/generated/model/debtPlanMethod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.30
 */

export type DebtPlanMethod = (typeof DebtPlanMethod)[keyof typeof DebtPlanMethod]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DebtPlanMethod = {
  avalanche: 'avalanche',
  snowball: 'snowball',
  custom: 'custom',
  none: 'none',
} as const